requirements.txt
apache-airflow==2.6.2
apache-airflow-providers-http==4.4.1
apache-airflow-providers-postgres==5.5.2
